Tidy dependency injection configuration
