+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/Infrastructure/Security/AuthenticationSchemeNames.cs
@@ -1,0 +1,7 @@
+namespace TeachingRecordSystem.Api.Infrastructure.Security;
+
+public static class AuthenticationSchemeNames
+{
+    public const string IdAccessToken = nameof(IdAccessToken);
+    public const string AuthorizeAccessAccessToken = nameof(AuthorizeAccessAccessToken);
+}
