+++ TeachingRecordSystem/src/TeachingRecordSystem.Core/DataStore/Postgres/Procedures/p_refresh_person_search_attributes_v8.sql
@@ -68,3 +68,5 @@ WHERE TRIM(COALESCE(attribute_value, '')) != '';
 
 END;
 $BODY$;
+
+
