+++ TeachingRecordSystem/src/TeachingRecordSystem.Core/Services/PersonMatching/ServiceCollectionExtensions.cs
@@ -7,6 +7,7 @@ public static class ServiceCollectionExtensions
     public static IServiceCollection AddPersonMatching(this IServiceCollection services)
     {
         services.AddTransient<IPersonMatchingService, PersonMatchingService>();
+
         return services;
     }
 }
