--- TeachingRecordSystem/src/TeachingRecordSystem.Core/Services/Webhooks/ApplicationBuilderExtensions.cs
@@ -1,28 +1,0 @@
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-
-namespace TeachingRecordSystem.Core.Services.Webhooks;
-
-public static class HostApplicationBuilderExtensions
-{
-    public static IHostApplicationBuilder AddWebhookOptions(this IHostApplicationBuilder builder)
-    {
-        builder.Services.AddOptions<WebhookOptions>()
-            .Bind(builder.Configuration.GetSection("Webhooks"))
-            .ValidateDataAnnotations()
-            .ValidateOnStart();
-
-        return builder;
-    }
-
-    public static IHostApplicationBuilder AddWebhookDeliveryService(this IHostApplicationBuilder builder)
-    {
-        AddWebhookOptions(builder);
-
-        WebhookSender.Register(builder.Services);
-
-        builder.Services.AddSingleton<IHostedService, WebhookDeliveryService>();
-
-        return builder;
-    }
-}
