--- TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Services/ServiceCollectionExtensions.cs
@@ -1,16 +1,0 @@
-using TeachingRecordSystem.SupportUi.Services.AzureActiveDirectory;
-
-namespace TeachingRecordSystem.SupportUi.Services;
-
-public static class ServiceCollectionExtensions
-{
-    public static IServiceCollection AddSupportUiServices(
-        this IServiceCollection services,
-        IConfiguration configuration,
-        IHostEnvironment environment)
-    {
-        services.AddAzureActiveDirectory(environment);
-
-        return services;
-    }
-}
