+++ TeachingRecordSystem/tests/TeachingRecordSystem.AuthorizeAccess.EndToEndTests/HostFixture.cs
@@ -87,6 +87,7 @@ public sealed class HostFixture(IConfiguration configuration) : IAsyncDisposable
 
                     services.Configure<OpenIddictServerAspNetCoreOptions>(options => options.DisableTransportSecurityRequirement = true);
                     services.AddSingleton<OneLoginCurrentUserProvider>();
+                    services.AddFakeXrm();
                     services.AddSingleton(
                         sp => ActivatorUtilities.CreateInstance<TestData>(sp, TestDataPersonDataSource.CrmAndTrs));
                     services.AddFakeXrm();
