+++ TeachingRecordSystem/tests/TeachingRecordSystem.Core.Tests/Startup.cs
@@ -1,7 +1,6 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Npgsql;
 using TeachingRecordSystem.Core.Services.Webhooks;
 using TeachingRecordSystem.Core.Tests.ApiSchema;
 using TeachingRecordSystem.Core.Tests.Jobs;
@@ -18,27 +17,21 @@ public class Startup
                 .AddEnvironmentVariables())
             .ConfigureServices((context, services) =>
             {
-                var pgConnectionString = new NpgsqlConnectionStringBuilder(context.Configuration.GetRequiredConnectionString("DefaultConnection"))
-                {
-                    // We rely on error details to get the offending duplicate key values in the TrsDataSyncHelper
-                    IncludeErrorDetail = true
-                }.ConnectionString;
+                DbHelper.ConfigureDbServices(services, context.Configuration.GetPostgresConnectionString());
 
-                DbHelper.ConfigureDbServices(services, pgConnectionString);
-
-                services.AddMemoryCache();
-
-                services.AddSingleton<DbFixture>();
-                services.AddSingleton<FakeTrnGenerator>();
-                services.AddFakeXrm();
-                services.AddSingleton<ReferenceDataCache>();
-                services.AddSingleton<WebhookReceiver>();
-                services.AddSingleton<PersonInfoCache>();
-                services.AddSingleton<EventMapperFixture>();
-                services.AddSingleton<NightlyEmailJobFixture>();
-                services.AddSingleton<IClock, TestableClock>();
-                services.AddSingleton<WebhookMessageFactory>();
-                services.AddSingleton<EventMapperRegistry>();
+                services
+                    .AddMemoryCache()
+                    .AddSingleton<DbFixture>()
+                    .AddSingleton<FakeTrnGenerator>()
+                    .AddFakeXrm()
+                    .AddSingleton<ReferenceDataCache>()
+                    .AddSingleton<WebhookReceiver>()
+                    .AddSingleton<PersonInfoCache>()
+                    .AddSingleton<EventMapperFixture>()
+                    .AddSingleton<NightlyEmailJobFixture>()
+                    .AddSingleton<IClock, TestableClock>()
+                    .AddSingleton<WebhookMessageFactory>()
+                    .AddSingleton<EventMapperRegistry>();
             });
