+++ TeachingRecordSystem/tests/TeachingRecordSystem.SupportUi.Tests/TestScopedServices.cs
@@ -1,3 +1,4 @@
+using Microsoft.Extensions.Options;
 using TeachingRecordSystem.Core.Services.Files;
 using TeachingRecordSystem.Core.Services.GetAnIdentityApi;
 using TeachingRecordSystem.Core.Services.TrnRequests;
@@ -29,6 +30,15 @@ public class TestScopedServices
         TrnRequestOptions = new TrnRequestOptions();
     }
 
+    public static void ConfigureServices(IServiceCollection services) =>
+        services
+            .AddTestScoped<IClock>(tss => tss.Clock)
+            .AddTestScoped(tss => tss.GetAnIdentityApiClientMock.Object)
+            .AddTestScoped(tss => tss.AzureActiveDirectoryUserServiceMock.Object)
+            .AddTestScoped<IFeatureProvider>(tss => tss.FeatureProvider)
+            .AddTestScoped(tss => tss.BlobStorageFileServiceMock.Object)
+            .AddTestScoped(tss => Options.Create(tss.TrnRequestOptions));
+
     public static TestScopedServices GetCurrent() =>
         _current.Value ?? throw new InvalidOperationException("No current instance has been set.");
 
@@ -55,4 +65,13 @@ public class TestScopedServices
     public Mock<IGetAnIdentityApiClient> GetAnIdentityApiClientMock { get; }
 
     public TrnRequestOptions TrnRequestOptions { get; }
+}
+
+file static class ServiceCollectionExtensions
+{
+    public static IServiceCollection AddTestScoped<T>(this IServiceCollection services, Func<TestScopedServices, T> resolveService)
+        where T : class
+    {
+        return services.AddTransient(_ => resolveService(TestScopedServices.GetCurrent()));
+    }
 }
